Factor out getPriorityEntriesHandler
